- Refactored densim code to use stl container classes and AdultCohort class. - Changed densim to use double instead of float for all floating point values. - Fixed bug where densim was switching the low and high ratio for double blood meal ratios. - Added verification code to ensure equivalence between old arrays and new containers. - Added Mosquito.h and Mosquit.cpp, holding AdultCohort class. - Fix order of declaration and usage for double blood meal paramters.

+++ trunk/simulation/dsport/Classes.h
@@ -12,22 +12,22 @@ namespace dsport {
 class SequentialInfectionValues
 {
 public:
-  float F1T2;                // first 1 then 2 - number of individuals
-  float F1T3;
-  float F1T4;
-  float F2T1;
-  float F2T3;
-  float F2T4;
-  float F3T1;
-  float F3T2;
-  float F3T4;
-  float F4T1;
-  float F4T2;
-  float F4T3;
-  float FMT1;
-  float FMT2;
-  float FMT3;
-  float FMT4;
+  double F1T2;                // first 1 then 2 - number of individuals
+  double F1T3;
+  double F1T4;
+  double F2T1;
+  double F2T3;
+  double F2T4;
+  double F3T1;
+  double F3T2;
+  double F3T4;
+  double F4T1;
+  double F4T2;
+  double F4T3;
+  double FMT1;
+  double FMT2;
+  double FMT3;
+  double FMT4;
 };
 
 
@@ -35,7 +35,7 @@ public:
 class VirusDesc           // virus parameters
 {
 public:
-  float Viremia;          // number of mosquito infectious doses
+  double Viremia;          // number of mosquito infectious doses
   int Incub;              // incubation period in days
   int Durat;              // duration of viremia in days
 };
@@ -54,31 +54,31 @@ public:
 class DebugOutput
 {
 public:
-  float CalcDeathsArraySize;
-  float CalcBirthsArraySize;
-  float InitInfectivesArraySize;
-  float DMealProp;
-  float EggersNew;
-  float BitersNew;
-  float EggersOld;
-  float BitersOld;
-  float EIPTranNew[4+1];
-  float BitersInfdNewDB[4+1];
-  float EIPTranOld[4+1];
-  float BitersInfdOldDB[4+1];
-  float BitersInfd[4+1];
-  float EggersInfv[4+1];
-  float BitersInfv[4+1];
-  float BitersTotal;
-  float MosqTotal;
-  float MosqInfvTotal[4+1];
-  float BitesPerPerson;
-  float MosqInocEstimate[4+1];
-  float NewDlyMosqInoc[4+1];
-  float NewInfd[4+1];
-  float OldInfd[4+1];
-  float HumInocEstimate[4+1];
-  float NewDlyHumInoc[4+1];
+  double CalcDeathsArraySize;
+  double CalcBirthsArraySize;
+  double InitInfectivesArraySize;
+  double DMealProp;
+  double EggersNew;
+  double BitersNew;
+  double EggersOld;
+  double BitersOld;
+  double EIPTranNew[4+1];
+  double BitersInfdNewDB[4+1];
+  double EIPTranOld[4+1];
+  double BitersInfdOldDB[4+1];
+  double BitersInfd[4+1];
+  double EggersInfv[4+1];
+  double BitersInfv[4+1];
+  double BitersTotal;
+  double MosqTotal;
+  double MosqInfvTotal[4+1];
+  double BitesPerPerson;
+  double MosqInocEstimate[4+1];
+  double NewDlyMosqInoc[4+1];
+  double NewInfd[4+1];
+  double OldInfd[4+1];
+  double HumInocEstimate[4+1];
+  double NewDlyHumInoc[4+1];
 };
 
 
@@ -86,21 +86,21 @@ public:
 class DiskSpooler
 {
 public:
-  float Incubate1;                        // Number of humans incubating
-  float Incubate2;                        // Number of humans incubating
-  float Incubate3;                        // Number of humans incubating
-  float Incubate4;                        // Number of humans incubating
-  float Viremic1;                         // Number of humans infective
-  float Viremic2;                         // Number of humans infective
-  float Viremic3;                         // Number of humans infective
-  float Viremic4;                         // Number of humans infective
-  float MosqTotal;                        // Total daily mosquitoes
-  float MosqInfvTotal[4+1];               // Total infective mosquitoes
+  double Incubate1;                        // Number of humans incubating
+  double Incubate2;                        // Number of humans incubating
+  double Incubate3;                        // Number of humans incubating
+  double Incubate4;                        // Number of humans incubating
+  double Viremic1;                         // Number of humans infective
+  double Viremic2;                         // Number of humans infective
+  double Viremic3;                         // Number of humans infective
+  double Viremic4;                         // Number of humans infective
+  double MosqTotal;                        // Total daily mosquitoes
+  double MosqInfvTotal[4+1];               // Total infective mosquitoes
   int InfvBites;                          // New potential inoculations
-  float EIPDevRate[4+1];                  // Extrinsic incubation period rate - daily
-  float NumHumans;                        // Number of humans
-  float SerPos[23+1][4+1];                // Percent of indiv. in age class that are seropositive
-  float NewHumCases[4+1];                 // Number of new human cases/day/serotype
+  double EIPDevRate[4+1];                  // Extrinsic incubation period rate - daily
+  double NumHumans;                        // Number of humans
+  double SerPos[23+1][4+1];                // Percent of indiv. in age class that are seropositive
+  double NewHumCases[4+1];                 // Number of new human cases/day/serotype
   SequentialInfectionValues SeqInfVals;   // Daily sequential value statistics
 
   DebugOutput _debugOutput;
